requirements.txt
psycopg2-binary==2.9.5
python-dotenv==1.0.0
kivy==2.1.0
kivymd==1.1.1